// src/image-analysis-service/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { Ai, Annotation, App, KvCache, Logger, MRNObject, ServiceStub, SmartSql, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  API_GATEWAY: ServiceStub<import('../api-gateway/index.js').default>;
  FASHION_DB: SmartSql;
  FRONTEND_SERVICE: ServiceStub<import('../frontend-service/index.js').default>;
  IMAGE_GENERATION_SERVICE: ServiceStub<import('../image-generation-service/index.js').default>;
  logger: Logger;
  mem: KvCache;
  ORDER_SERVICE: ServiceStub<import('../order-service/index.js').default>;
  PAYMENT_SERVICE: ServiceStub<import('../payment-service/index.js').default>;
  PRODUCT_SERVICE: ServiceStub<import('../product-service/index.js').default>;
  tracer: Tracer;
  VULTR_API_KEY: string;
}